fix(context-queries): address code review issues on Plan C Task 1 queries
- Rename `trigger_user_id` → `_trigger_user_id` to signal intentional
  unused parameter; docstring says "reserved for future filtering"
- Add `.limit(max_same_chat + max_other_chat * 20)` to prevent pulling
  hundreds of rows in a busy 4h window before in-memory partition drops extras
- Extract `ScalarResult`/`IterResult` test fakes into `tests/unit/data/_helpers.py`
  to eliminate the duplicate class definitions in test_v4_queries.py and
  test_context_queries.py
- Fix UP017 ruff lint: `timezone.utc` → `UTC` (imported directly); fix
  import sort order in test_context_queries.py
