+++ apps/agent-service/tests/unit/data/_helpers.py
@@ -1,0 +1,25 @@
+"""Shared test fakes for unit tests that mock AsyncSession.execute results."""
+
+from __future__ import annotations
+
+
+class ScalarResult:
+    def __init__(self, value):
+        self.value = value
+
+    def scalar_one(self):
+        return self.value
+
+    def scalar_one_or_none(self):
+        return self.value
+
+
+class IterResult:
+    def __init__(self, values):
+        self._values = values
+
+    def scalars(self):
+        return self
+
+    def all(self):
+        return self._values
